fix: ssr构建异常

+++ src/plugins/ssr/index.js
@@ -1,5 +1,6 @@
 const fs = require('fs-extra');
 const path = require('path');
+const { warn } = require('../../utils/logger');
 
 function deleteLoading(str) {
   const s = str.substring(str.indexOf('<!--CLIENT_ONLY_START-->'), str.lastIndexOf('<!--CLIENT_ONLY_END-->'));
@@ -18,8 +19,13 @@ class SSR {
         const file = `${pageName[pageName.length - 2]}.html`;
         const srcFile = path.join(distDir, file);
         const targetFile = path.join(viewDir, file);
-        const data = deleteLoading(fs.readFileSync(srcFile).toString());
-        fs.writeFileSync(targetFile, data);
+        if (fs.existsSync(srcFile)) {
+          const data = deleteLoading(fs.readFileSync(srcFile).toString());
+          fs.writeFileSync(targetFile, data);
+        } else {
+          console.log();
+          warn(`ssr entry "${key}" not found html file!`);
+        }
       });
       callback();
     });
